requirements.txt
openenv-core>=0.2.2
openai>=1.0.0
google-genai>=0.1.0
fastapi
uvicorn
pydantic
